# Optimize PathGradientBrush, cleanup InternalPath, share ThreadLocal code