… .github/workflows/build-and-test.yml
… src/ImageSharp.Drawing/Processing/SolidBrush.cs
… src/ImageSharp.Drawing/Shapes/ComplexPolygon.cs
… src/ImageSharp.Drawing/Shapes/IPath.cs
… src/ImageSharp.Drawing/Shapes/InternalPath.cs
… src/ImageSharp.Drawing/Shapes/Path.cs
… src/ImageSharp.Drawing/Shapes/RectangularPolygon.cs
… src/ImageSharp.Drawing/Utilities/Intersect.cs
… src/ImageSharp.Drawing/Utilities/ThreadLocalBlenderBuffers.cs
… tests/ImageSharp.Drawing.Benchmarks/Drawing/FillPathGradientBrush.cs
… tests/ImageSharp.Drawing.Tests/Drawing/FillPathGradientBrushTests.cs
… tests/ImageSharp.Drawing.Tests/Shapes/InternalPathExtensions.cs
… tests/ImageSharp.Drawing.Tests/Shapes/InternalPathTests.cs
… tests/ImageSharp.Drawing.Tests/Shapes/Issues/Issue_16.cs
… tests/ImageSharp.Drawing.Tests/Shapes/Issues/Issue_19.cs
… tests/ImageSharp.Drawing.Tests/Shapes/Issues/Issue_ClippedPaths.cs
… tests/ImageSharp.Drawing.Tests/Shapes/PolygonTests.cs
… tests/ImageSharp.Drawing.Tests/Shapes/RectangleTests.cs
… tests/ImageSharp.Drawing.Tests/TestUtilities/ApproximateFloatComparer.cs
… tests/ImageSharp.Drawing.Tests/TestUtilities/TestMemoryAllocator.cs
… tests/ImageSharp.Drawing.Tests/Utilities/IntersectTests.cs
… tests/ImageSharp.Drawing.Tests/Utilities/ThreadLocalBlenderBuffersTests.cs