### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp.Drawing/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Follow up on https://github.com/SixLabors/ImageSharp.Drawing/pull/128#issuecomment-827738628 and https://github.com/SixLabors/ImageSharp.Drawing/pull/128#discussion_r622002284:
- Simplify and speed up `PathGradientBrush` with basic intersection code. (~5 speedup on a basic benchmark)
- Remove `IPath.Contains`. Although did not mention in #103 it also solves computational geometry concern which is unscoped, and current implementation relied on `InternalPath.FindIntersections`
- Remove `InternalPath.FindIntersections` and all related code
- Shared utility for thread local blender buffers

#### Benchmarks

Before:
```
|                       Method |     Mean |     Error |   StdDev |   Median |
|----------------------------- |---------:|----------:|---------:|---------:|
| FillGradientBrush_ImageSharp | 9.726 ms | 0.3696 ms | 1.030 ms | 9.368 ms |
```

After:
```
|                       Method |     Mean |     Error |    StdDev |
|----------------------------- |---------:|----------:|----------:|
| FillGradientBrush_ImageSharp | 2.144 ms | 0.0769 ms | 0.2206 ms |
```