@antonfirsov 2021-04-30
ImageMagick hates my new reference image:
https://github.com/SixLabors/ImageSharp.Drawing/blob/0da47ef2ee5b3772c48324ecdaf01178751f8cd8/tests/Images/ReferenceOutput/Drawing/GradientBrushes/FillPathGradientBrushTests/FillComplex.png

```
Error Message:
   ImageMagick.MagickMissingDelegateErrorException : NoDecodeDelegateForThisImageFormat `' @ error/blob.c/CustomStreamToImage/804
```

@dlemstra any idea? Is the version up-to-date in this repo?

https://github.com/SixLabors/ImageSharp.Drawing/blob/d8960b8e2e55875d2c4f990f5fc439b846af1230/tests/Directory.Build.targets#L22