@JimBobSquarePants 2021-04-30
@antonfirsov It'll be LFS. Looks like new images are being automatically added due to the new editorconfig. 

You need to enable LFS in the workflow and also enable caching. 

Copy the yml from ImageSharp, update the infrastructure submodule, and change the `XUNIT_PATH` path variable and all should be ok. 

https://github.com/SixLabors/ImageSharp/blob/a6c17854684edaa998080233273d19d987174bb5/.github/workflows/build-and-test.yml

I can do this if you want for ease?