@antonfirsov 2021-04-30
> Update build-and-test.yml bc50f77

Thanks! 

I'll leave this open for a while to give @tocsoft a chance to react, if he has the time.